+++ Electrode/src/Electrode_RadialDiffRegions.h
@@ -194,8 +194,7 @@ public:
 
 
     //! Get all of the reaction rates and parameters from Cantera
-    void extractInfoJustBorn(std::vector<int>& justBornMultiSpecies);
-
+    void extractInfoJustBorn(std::vector<size_t>& justBornMultiSpecies);
 
     //! Print conditions of the electrode for the current integration step to stdout
     /*!
