+++ Electrode/src/Electrode_SimpleDiffCVode.h
@@ -52,7 +52,8 @@ public:
     int electrode_model_create(ELECTRODE_KEY_INPUT* ei);
 
     void calcRate(double deltaT);
-    void extractInfo(std::vector<int>& justBornMultiSpecies);
+
+    void extractInfo(std::vector<size_t>& justBornMultiSpecies);
 
 
     //! Print conditions of the electrode for the current integration step to stdout
